refactor: 행정구역별 상세 정보 리스트 조회 CategoryController로 이동
- API 경로 및 메서드 명, DTO 명 수정
- CategoryController 정리
