+++ src/main/java/com/ssafy/firskorea/attraction/dto/response/category/SidoDetailDto.java
@@ -1,19 +1,15 @@
-package com.ssafy.firskorea.plan.dto;
+package com.ssafy.firskorea.attraction.dto.response.category;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.ToString;
 
 @Getter
-@Setter
 @NoArgsConstructor
 @ToString
-public class RegionDto {
+public class SidoDetailDto {
     private String sidoCode;
     private String sidoName;
-//    private String gugunName;
-//    private String gugunCode;
     private String sidoDescription;
     private String sidoImage;
 }
