+++ src/main/java/com/ssafy/firskorea/attraction/service/AttractionService.java
@@ -10,6 +10,7 @@ import com.ssafy.firskorea.attraction.dto.request.SearchDto;
 import com.ssafy.firskorea.attraction.dto.request.SidoPgnoDto;
 import com.ssafy.firskorea.attraction.dto.response.AttractionDto;
 import com.ssafy.firskorea.attraction.dto.response.PaginatedAttractionsDto;
+import com.ssafy.firskorea.attraction.dto.response.category.SidoDetailDto;
 import com.ssafy.firskorea.attraction.dto.response.category.ThemeDto;
 
 public interface AttractionService {
@@ -19,6 +20,8 @@ public interface AttractionService {
     List<Category> getCategoryList(Character code) throws SQLException;
 
     List<Category> getSidoList() throws SQLException;
+
+    List<SidoDetailDto> getSidoInfoList() throws SQLException;
 
     List<AttractionDto> getAttractionsBySearch(SearchDto searchDto) throws SQLException;
